+++ components/st-radio-menu/index.ts
@@ -1,0 +1,1 @@
+export * from './st-radio-menu.component';