+++ web/app/+examples/index.ts
@@ -13,3 +13,4 @@ export * from './regexp';
 export * from './horizontal-tab';
 export * from './tab-box';
 export * from './api-doc';
+export * from './radio-menu';